requirements.txt
youtube_dl
discord
python-dotenv
ffmpeg-python
psycopg2